[nw] same speed as CPU now

+++ src/cuda/cudabatch.hpp
@@ -130,7 +130,7 @@ protected:
     uint16_t * sequence_lengths_d_;
 
     // Buffers for temp data.
-    int32_t* scores_d_;
+    int16_t* scores_d_;
     int16_t* traceback_i_d_;
     int16_t* traceback_j_d_;
